Minor changes
made some changes to sort the printing of the listing of outstanding orders

+++ .code/Tab.h
@@ -54,7 +54,7 @@ public:
      * @brief Get the total price of outstanding orders in the tab.
      * @return The total price of outstanding orders.
      */
-    int getTotal();
+    float getTotal();
 
     void setDescription(string d);
